Make mapperconnection a singleton, change test names
mapperconnection is changed to a singleton so only one sqlalchemy engine
is being used. This is the most efficient according to docs. Changed
the test to initialise that mapperconnection to the test_db. Also
changed the names of the test data assets in case they accidentally end
up in the real db. It will be easier to distinguish them.